updates to change App.js to index.jsx and others to get summary to work in project structure
