+++ react-client/src/index.jsx
@@ -1,36 +1,60 @@
-import React from 'react';
+import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
-import $ from 'jquery';
-import List from './components/List.jsx';
+import BillSplits from './components/BillSplits'
+import UserInfo from './components/UserInfo'
+import DebtsSummary from './components/DebtsSummary'
+//import './index.css';
 
-class App extends React.Component {
+class App extends Component {
   constructor(props) {
     super(props);
-    this.state = { 
-      items: []
+    this.state = {
+      billsplititems : []
     }
   }
 
-  componentDidMount() {
-    $.ajax({
-      url: '/items', 
-      success: (data) => {
-        this.setState({
-          items: data
-        })
+  componentWillMount() {
+    this.setState({
+      billsplititems : [{
+        "event":"Sushi Dinner",
+        "debtor":"Robert",
+        "creditor":"Group1",
+        "eventid":"174",
+        "amount":"25.50"
       },
-      error: (err) => {
-        console.log('err', err);
+      {
+        "event":"Ski Trip Tahoe",
+        "debtor":"Eve",
+        "creditor":"Group2",
+        "eventid":"178",
+        "amount":"27.00"
+      },
+      {
+        "event":"BBQ Picnic",
+        "debtor":"Zander",
+        "creditor":"Group3",
+        "eventid":"179",
+        "amount":"56.00"
       }
-    });
+    ]})
   }
 
-  render () {
-    return (<div>
-      <h1>Item List</h1>
-      <List items={this.state.items}/>
-    </div>)
+  render() {
+    return (
+      <div className="App">
+      <header>BillSplits</header>
+      <br/>
+      <UserInfo />
+      <p></p>
+      <DebtsSummary />
+      <br />
+      <br />
+      <splitbutton><a>Add a New Split</a></splitbutton>
+            <br />
+            <br />
+        <BillSplits billsplititems={this.state.billsplititems} />
+      </div>
+    );
   }
 }
-
 ReactDOM.render(<App />, document.getElementById('app'));
